+++ starsight/src/marks.rs
@@ -4,6 +4,6 @@
 //! mark types ready to be added to a [`Figure`](crate::figures::Figure).
 
 pub use crate::components::marks::{
-    BarMark, DataExtent, HeatmapMark, HistogramMark, LineMark, Mark, PointMark, StepMark,
-    StepPosition,
+    AreaBaseline, AreaMark, BarMark, BarRenderContext, DataExtent, HeatmapMark, HistogramMark,
+    LineMark, Mark, Orientation, PointMark, StepMark, StepPosition,
 };